Implement a basic ray tracer
An extremely basic ray tracer has been implemented that can be built on
top of. This ray tracer simply uses 1 ray per pixel and outputs depth
information, clamped such that a depth of 0 is entirely white and a
depth of 5 is entirely black.

Additionally, a simplistic image printer has been added that uses xterm
truecolor escape codes to render the image (in relatively low quality)
to the terminal.

Currently, this is tested using a simplistic, hard-coded scene with 2
spheres of different sizes placed such that the smaller one partially
obscures the larger one.