Implement a dummy BVH
In the future, it will likely be useful to use a bounding volume
hierarchy to speed up ray intersections. To allow this to be easily
added in the future, a dummy BVH<T> class has been added which can be
used for this purpose.

The BVH<T> class itself is fully functional, but the BVH<T>::construct
static method used to construct the BVH is currently a simple dummy
function that creates a BVH with one node containing all of the objects.