Implement PPM image output
The result of the ray tracer can now be saved to a plain PPM image. The
default behaviour of the ray tracer is to render the scene twice: once
at low resolution for the console preview and once at a higher
resolution for saving to the PPM file, which is currently hardcoded to
be saved as "test.ppm".