Implement supersampling
To achieve more pleasing results, it is common for ray tracing
algorithms to cast multiple rays per pixel and use the average of their
colours as the colour of the pixel; this is called supersampling.

The ray trace renderer now supports this by providing a "supersample
level", which is just the number of samples to take in each of the x and
y directions. So the total number of samples taken is the square of this
value.