Add progress indicators
Previously, there were no indicators of progress besides seeing the
preview show up. This may pose a problem in the future when rendering
more complex scenes, since these will take a while. To indicate progress
to the user, a spinner is now shown during all stages and a progress
percentage is now shown when performing the actual renders.