Remove backface culling
Previously, both the triangle and sphere algorithms for determining
intersections were designed to not register an intersection with
back-facing primitives. However, this will cause problems when
transmissive materials are implemented, since backfaces must be
considered for that to work. It was also causing problems for shadow
rays under certain conditions.

Because of these issues, backface culling has been removed entirely and
intersections with backfaces will now be registered.