Implement BVH construction
Previously, BVH construction was dummied out and all BVHs would be
comprised of a single node containing all elements. This has been
replaced with proper BVH construction using approximate agglomerative
clustering.

In essence, when trying to split a large number of elements, the split
is done based on the morton code of the centroid of each element; this
is very fast but produces non-optimal results. When trying to split a
small number of elements, a slower but better greedy clustering
algorithm is used to create nodes. Together, this results in a
reasonably efficient BVH with a time complexity of O(N log N). Using the
greedy algorithm for the entire BVH construction would have a time
complexity of O(N^3), which would be unacceptable.

Additionally, constructing a BVH per triangulated mesh provides several
important benefits. The most obvious benefit of doing so is that having
multiple copies of the same object in the scene does not noticably
increase the time taken to perform BVH construction. On top of that,
this also seriously decreases the memory footprint of the BVHs that need
to be searched. While memory might not seem like a major issue, such a
reduction in memory footprint can actually result in significantly
improved cache performance for many scenes.