Implement reading command line arguments and scene files
Previously, almost everything about the test scene and the rendering
method was hardcoded. However, it is necessary in the future to be able
to configure those parameters without recompiling. In order to do this,
command line arguments need to be read and parsed. This has been
implemented.

On top of that, the scene loader from HW3 has been adapted to work with
this program. A couple of minor changes had to be made, but it works the
same overall. This should allow the chessboard scene from HW3 to be
ported over quite easily.