Add a chessboard scene
Previously, the only available scene was the knight test scene. A more
complex scene is required to show off all the features of the ray
tracer. In order to do this, the chessboard scene from HW3 has been
ported to work with the ray tracer.

This scene is very complicated, containing 9 different triangular
meshes, each with tens of thousands of triangles. Not only does this
scene provide ample opportunity to show off the various different
features of the ray tracer (shadows, specular rays, etc.), it also shows
off just how performant the ray tracer really is.

Since the intention is to use this scene as the demonstration scene for
the assignment, the Python wrapper has been modified to use it.