Remove unneeded shared_ptr in Intersection
Previously, the material returned by an intersection was being stored
using a shared_ptr<Material>. While this is correct, it was causing a
small but noticeable performance hit since every time a shared_ptr is
copied, an atomic operation needs to be performed on its refcount.

This performance hit was not necessary, since it is guaranteed that the
material is also referred to by the Object itself, which should be alive
for the entire lifetime of any Intersection it creates. Thus, replacing
this with a raw pointer can avoid unneeded atomic operations, reducing
contention between threads and slightly improving performance.