Handle total internal reflection
Previously, the plan was to add total internal reflection when adding
simulation of the Fresnel equations. However, this would require keeping
track of the current IOR the ray is passing through which is simply too
much work to add this close to the deadline.

Instead, total internal reflection is now handled as a special case when
glm::refract returns (NaN, NaN, NaN).