Fix BoundingBox::intersects returning true when it shouldn't
Previously, there was a minor issue where the BoundingBox::intersects
function could return true even if the given ray does not actually
intersect the box. This was happening when the ray intersects properly
in both the x and y directions and intersects backwards in the z
direction.

This was caused by the tmax variable not being updated using tzmax. In a
previous version of the program, this was not an issue since
intersections in the -t direction were considered valid. However, when
the tmax > 0 check was added, the code to update tmax was not added.

This problem did not cause any incorrect output, since bounding box
intersections are only used for acceleration. Instead, the issue was
causing BVH nodes to be searched when they didn't need to be in scenes
with a lot of geometry behind the camera when the camera was oriented in
a particular way.

In fixing this problem, the AABB intersection algorithm has been
replaced with a branchless algorithm, which is both clearer and provides
a marginal performance improvement.